refactor: standardize middleware and error handling
